requirements.txt
APScheduler
Pyrogram
pytz
SQLAlchemy
TgCrypto
python-dotenv
Flask
